requirements.txt
South==0.7.3
Werkzeug==0.8.3
coverage==3.5.1
django-debug-toolbar==0.9.4
django-extensions==0.8
django-nose==1.0
-e git+git@github.com:tangentlabs/django-oscar.git#egg=django_oscar
mock==0.7.2
nose==1.1.2
